+++ src/lte/model/pss-ff-mac-scheduler.cc
@@ -1392,7 +1392,7 @@ PssFfMacScheduler::DoSchedDlTriggerReq (const struct FfMacSchedSapProvider::Sche
       if (lcActives == 0)
         {
           // Set to max value, to avoid divide by 0 below
-          lcActives = 65535; // UINT16_MAX;
+          lcActives = (uint16_t)65535; // UINT16_MAX;
         }
       uint16_t RgbPerRnti = (*itMap).second.size ();
       std::map <uint16_t,SbMeasResult_s>::iterator itCqi;
